Refactor the Builders based on Sonar warnings.
