+++ src/main/java/tfw/tsm/Initiator.java
@@ -278,4 +278,40 @@ public class Initiator extends TreeComponent {
             this.setLocation = TransactionMgr.isTraceLogging() ? new Throwable("StateChange") : null;
         }
     }
+
+    public static class InitiatorBuilder {
+        private String name = null;
+        private List<EventChannelDescription> eventChannelDescriptions = new ArrayList<>();
+        private StateQueueFactory stateQueueFactory = new DefaultStateQueueFactory();
+
+        InitiatorBuilder() {}
+
+        public InitiatorBuilder setName(final String name) {
+            this.name = name;
+
+            return this;
+        }
+
+        public InitiatorBuilder addEventChannelDescription(final EventChannelDescription eventChannelDescription) {
+            eventChannelDescriptions.add(eventChannelDescription);
+
+            return this;
+        }
+
+        public InitiatorBuilder setStateQueueFactory(final StateQueueFactory stateQueueFactory) {
+            this.stateQueueFactory = stateQueueFactory;
+
+            return this;
+        }
+
+        public Initiator build() {
+            Argument.assertNotNull(name, "name");
+            Argument.assertGreaterThan(eventChannelDescriptions.size(), 0, "eventChannelDescriptions.size()");
+
+            final EventChannelDescription[] ecds =
+                    eventChannelDescriptions.toArray(new EventChannelDescription[eventChannelDescriptions.size()]);
+
+            return new Initiator(name, ecds, stateQueueFactory);
+        }
+    }
 }
